// client/src/state/auth/types.ts
// Copyright 2019 Authors of Hubble
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//     http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.
export interface StatusState {
  readonly loading: boolean;
  readonly authenticated: boolean;
  readonly error?: any;
}

export interface UserState {}

export interface State {
  readonly status: StatusState;
  readonly user: UserState;
  readonly token: string | null;
}